Root Eternal
After never managing to achieve the pass degree in calculus despite being a literal robot. 

Our hero decides to solve the problem the hard way .

GUNS .

shoot the root out of your enemies and engage in intense situations against the numbers in this top down rooter shooter .

Shoot and root , until it's done. 

Unless of course you are 2^2